backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
SQLAlchemy==2.0.31
pymongo==4.8.0
sqlglot==25.6.0
pymysql==1.1.1
requests==2.32.3
torch>=2.6.0
sentence-transformers>=2.7.0
transformers>=4.40.0
spacy>=3.7.0
